+++ src/provider/client/types/vmpc/index.ts
@@ -1,2 +1,3 @@
 export * from './listHostDetections';
 export * from './listVulnerabilities';
+export * from './listScannedHostIds';